requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
anthropic==0.40.0
gspread==6.1.2
google-auth==2.35.0
pydantic==2.9.0
jinja2==3.1.4
python-dotenv==1.0.1
cachetools==5.5.0
httpx==0.27.0
pytest==8.3.0
